fix: Fixed

+++ src/Modules/Tours/Explorer.Tours.Core/Domain/KeyPoint.cs
@@ -8,9 +8,9 @@ public class KeyPoint : Entity {
     public string Name { get; init; }
     public string? Description { get; init; }
     public byte[]? Image { get; init; }
-    public int TourId { get; init; }
+    public long TourId { get; init; }
 
-    public KeyPoint(string name, string? description, double latitude, double longitude, byte[]? image, int tourId) {
+    public KeyPoint(string name, string? description, double latitude, double longitude, byte[]? image, long tourId) {
         if (string.IsNullOrWhiteSpace(name)) 
             throw new ArgumentException("Invalid Name.");
         Latitude = latitude;
